+++ ReleaseNotes.md
@@ -2,8 +2,12 @@
 - Обнаружение плоскостей в облаке точек
 - Возможность выключить/включить логирование в файл
 - Возможность передавать данные на другие запущенные экземпляры приложения
+- Возможность менять прозрачность линий
+- Камера может следовать за отслеживаемым объектом
 
 # New in this version:
 - Planes detection tool
 - User can disable/enable logging in file
 - Plugin that allows Elektronik to transmit data from one instance to another
+- Semitransparent lines
+- Camera can follow tracked objects